apps/webapp/serializers.py
from rest_framework import serializers
from .models import Task

class TaskSerialzier(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = '__all__'


class CreateTaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ('title', 'description', )